Rename indexing -> indexed
